requirements.txt
appnope==0.1.4
asttokens==3.0.0
certifi==2025.1.31
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
gensim==0.10.1
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
scipy==1.16.2
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
wcwidth==0.2.14
pandas==2.3.2
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.3.0
scikit-learn==1.7.2
openai==2.0.0
python-dotenv==1.1.1
nilearn==0.12.1
neuroquery==1.1.0
